% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_design_BayDesign.R
\name{BayDesign}
\alias{BayDesign}
\title{Initialization function for "Design"}
\usage{
BayDesign(model, stopping, increments, PLcohortSize = CohortSizeConst(0L), ...)
}
\arguments{
\item{model}{see \code{\linkS4class{Design}}}

\item{stopping}{see \code{\linkS4class{Design}}}

\item{increments}{see \code{\linkS4class{Design}}}

\item{PLcohortSize}{see \code{\linkS4class{Design}}}

\item{\dots}{additional arguments for \code{\link{RuleDesign}}}
}
\value{
the \code{\linkS4class{BayDesign}} object
}
\description{
Initialization function for "Design"
}
\keyword{methods}
