% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_design_1GeneralSimulationsBay_summary.R
\name{GeneralSimulationsBay}
\alias{GeneralSimulationsBay}
\title{Initialization function for "GeneralSimulations"}
\usage{
GeneralSimulationsBay(data, doses, seed)
}
\arguments{
\item{data}{see \code{\linkS4class{GeneralSimulations}}}

\item{doses}{see \code{\linkS4class{GeneralSimulations}}}

\item{seed}{see \code{\linkS4class{GeneralSimulations}}}
}
\value{
the \code{\linkS4class{GeneralSimulations}} object
}
\description{
Initialization function for "GeneralSimulations"
}
\keyword{methods}
