% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction.R
\name{IncrementMinDLT1N1}
\alias{IncrementMinDLT1N1}
\title{Initialization function for "IncrementMinDLT1N1"}
\usage{
IncrementMinDLT1N1(IncrementsList)
}
\arguments{
\item{IncrementsList}{see \code{\linkS4class{IncrementMinDLT1N1}}}
}
\value{
the \code{\linkS4class{IncrementMinDLT1N1}} object
}
\description{
Initialization function for "IncrementMinDLT1N1"
}
\keyword{methods}
