% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_IncrementsAbsoluteMax.R
\name{IncrementsAbsoluteMax}
\alias{IncrementsAbsoluteMax}
\title{Initialization function for "IncrementsAbsoluteMax"}
\usage{
IncrementsAbsoluteMax(intervals, absoluteIncrements)
}
\arguments{
\item{intervals}{see \code{\linkS4class{IncrementsAbsoluteMax}}}

\item{absoluteIncrements}{see \code{\linkS4class{IncrementsAbsoluteMax}}}
}
\value{
the \code{\linkS4class{IncrementsAbsoluteMax}} object
}
\description{
Initialization function for "IncrementsAbsoluteMax"
}
\keyword{methods}
