% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_IncrementsHSRBeta_DK.R,
%   R/x_new_IncrementsHSRbeta.R
\docType{class}
\name{IncrementsHSRBeta-class}
\alias{IncrementsHSRBeta-class}
\alias{.IncrementsHSRBeta}
\alias{IncrementsHSRBeta}
\title{Increments control based on eligable dose levels according to Hard Safety
Rule - Beta(a,b)}
\usage{
IncrementsHSRBeta(target = 0.3, prob = 0.95, a = 1, b = 1)
}
\arguments{
\item{target}{(\code{number})\cr the target toxicity}

\item{prob}{(\code{number})\cr the threshold probability for a dose being toxic}

\item{a}{(\code{number})\cr shape parameter a>0 of probability
distribution Beta (a,b)}

\item{b}{(\code{number})\cr shape parameter b>0 of probability
distribution Beta (a,b)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{IncrementsHSRBeta}} is a class for increments based on the Bin-Beta model.
Increment control is based on the number of observed DLTs and number of
subjects at each dose level. The probability of toxicity is calculated
using a Bin-Beta model with prior (a,b). If the probability exceeds
the threshold for a given dose, that dose and all doses above are excluded
from further escalation.
This is a hard safety rule that limits further escalation based on the
observed data per dose level.
}
\section{Slots}{

\describe{
\item{\code{target}}{a probability > 0 and < 1 of the target toxicity rate}

\item{\code{prob}}{a probability > 0 and < 1 of the confidence of the test}

\item{\code{a}}{positive real number representing the shape parameter of a Beta(a,b) distribution}

\item{\code{b}}{positive real number representing the shape parameter of a Beta(a,b) distribution}

\item{\code{target}}{(\code{number})\cr the target toxicity}

\item{\code{prob}}{(\code{number})\cr the threshold probability for a dose being toxic}

\item{\code{a}}{(\code{number})\cr shape parameter a>0 of probability
distribution Beta (a,b)}

\item{\code{b}}{(\code{number})\cr shape parameter b>0 of probability
distribution Beta (a,b)}
}}

\examples{

# Limit the escalation with a hard safety criteria to the doses that are below
# the first dose that is toxic with a probability of 0.95
my_increments <- IncrementsHSRBeta(target = 0.3, prob = 0.95)
}
\keyword{classes}
