% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_IncrementsHSRBeta_DK.R
\name{IncrementsHSRBeta}
\alias{IncrementsHSRBeta}
\title{Initialization function for "IncrementsHSRBeta"}
\usage{
IncrementsHSRBeta(target = 0.3, prob = 0.95, a = 1, b = 1)
}
\arguments{
\item{target}{see \code{\linkS4class{IncrementsHSRBeta}}}

\item{prob}{see \code{\linkS4class{IncrementsHSRBeta}}}

\item{a}{see \code{\linkS4class{IncrementsHSRBeta}}}

\item{b}{see \code{\linkS4class{IncrementsHSRBeta}}}
}
\value{
the \code{\linkS4class{IncrementsHSRBeta}} object
}
\description{
Initialization function for "IncrementsHSRBeta"
}
\keyword{methods}
