% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_IncrementsHSRBeta_DK.R
\docType{class}
\name{IncrementsHSRBeta_dk-class}
\alias{IncrementsHSRBeta_dk-class}
\alias{.IncrementsHSRBeta_dk}
\title{Increments control based on eligable dose levels according to Hard Safety
Rule - Beta(a,b)}
\description{
Increments control based on eligable dose levels according to Hard Safety
Rule - Beta(a,b)
}
\section{Slots}{

\describe{
\item{\code{target}}{a probability > 0 and < 1 of the target toxicity rate}

\item{\code{prob}}{a probability > 0 and < 1 of the confidence of the test}

\item{\code{a}}{positive real number representing the shape parameter of a Beta(a,b) distribution}

\item{\code{b}}{positive real number representing the shape parameter of a Beta(a,b) distribution}
}}

\keyword{classes}
