% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_IncrementsHSRBeta_DK.R
\name{IncrementsHSRBeta_dk}
\alias{IncrementsHSRBeta_dk}
\title{Initialization function for "IncrementsHSRBeta_dk"}
\usage{
IncrementsHSRBeta_dk(target, prob, a, b)
}
\arguments{
\item{target}{see \code{\linkS4class{IncrementsHSRBeta_dk}}}

\item{prob}{see \code{\linkS4class{IncrementsHSRBeta_dk}}}

\item{a}{see \code{\linkS4class{IncrementsHSRBeta_dk}}}

\item{b}{see \code{\linkS4class{IncrementsHSRBeta_dk}}}
}
\value{
the \code{\linkS4class{IncrementsHSRBeta_dk}} object
}
\description{
Initialization function for "IncrementsHSRBeta_dk"
}
\keyword{methods}
