% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction.R
\docType{class}
\name{IncrementsNlastDL-class}
\alias{IncrementsNlastDL-class}
\alias{.IncrementsNlastDL}
\title{Increments control based on relative differences in terms of DLTs}
\description{
Note that \code{DLTintervals} is to be read as follows. If for example,
we want to specify three intervals: First 0 DLTs, second 1 or 2 DLTs, and
third at least 3 DLTs, then we specify
\code{DLTintervals} to be \code{c(0, 1, 3)}. That means, the right
bound of the intervals are exclusive to the interval -- the vector only
gives the left bounds of the intervals. The last interval goes from 3 to
infinity.
}
\section{Slots}{

\describe{
\item{\code{NlastDL}}{an integer with the min number of patients in case that a DLT occurs}

\item{\code{DLTlastDL}}{an integer with the number of DLTs where an expansion should happen
relative increments in the \code{DLTintervals}}
}}

\keyword{classes}
