% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction.R
\name{IncrementsNlastDL}
\alias{IncrementsNlastDL}
\title{Initialization function for "IncrementsRelativeDLT"}
\usage{
IncrementsNlastDL(NlastDL, DLTlastDL)
}
\arguments{
\item{NlastDL}{see \code{\linkS4class{IncrementsNlastDL}}}

\item{DLTlastDL}{see \code{\linkS4class{IncrementsNlastDL}}}
}
\value{
the \code{\linkS4class{IncrementsNlastDL}} object
}
\description{
Initialization function for "IncrementsRelativeDLT"
}
\keyword{methods}
