% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\docType{class}
\name{IncrementsNumDoseLevelsBaySafetyFix-class}
\alias{IncrementsNumDoseLevelsBaySafetyFix-class}
\alias{.IncrementsNumDoseLevelsBaySafetyFix}
\title{Increments control based the maximum dose level increase
(not relative to the dose, absolute to the number of the dose level)
and a rule spefied in the protocol}
\description{
Increments control based the maximum dose level increase
(not relative to the dose, absolute to the number of the dose level)
and a rule spefied in the protocol
}
\section{Slots}{

\describe{
\item{\code{maxLevels}}{a vector with the left bounds of the relevant intervals}

\item{\code{toxrule}}{the toxrule for lower doses only}

\item{\code{toxrulesame}}{toxrule for lower and the same dose dose}
}}

\keyword{classes}
