% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\name{IncrementsNumDoseLevelsBaySafetyFix}
\alias{IncrementsNumDoseLevelsBaySafetyFix}
\title{Initialization function for "IncrementsRelativeSafetyFix"}
\usage{
IncrementsNumDoseLevelsBaySafetyFix(maxLevels = 1, toxrule, toxrulesame)
}
\arguments{
\item{maxLevels}{see \code{\linkS4class{IncrementsNumDoseLevelsBaySafetyFix}}}

\item{toxrule}{see \code{\linkS4class{IncrementsNumDoseLevelsBaySafetyFix}}}

\item{toxrulesame}{see \code{\linkS4class{IncrementsNumDoseLevelsBaySafetyFix}}}
}
\description{
Initialization function for "IncrementsRelativeSafetyFix"
}
\keyword{methods}
