% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_IncrementsRelativeDLTmax.R
\name{IncrementsRelativeDLTmax}
\alias{IncrementsRelativeDLTmax}
\title{Initialization function for "IncrementsRelativeDLTmax"}
\usage{
IncrementsRelativeDLTmax(DLTintervals, increments)
}
\arguments{
\item{DLTintervals}{see \code{\linkS4class{IncrementsRelativeDLTmax}}}

\item{increments}{see \code{\linkS4class{IncrementsRelativeDLTmax}}}
}
\value{
the \code{\linkS4class{IncrementsRelativeDLTmax}} object
}
\description{
Initialization function for "IncrementsRelativeDLTmax"
}
\keyword{methods}
