% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_IncrementsRelativeMax.R
\name{IncrementsRelativeMax}
\alias{IncrementsRelativeMax}
\title{Initialization function for "IncrementsRelative"}
\usage{
IncrementsRelativeMax(intervals, increments)
}
\arguments{
\item{intervals}{see \code{\linkS4class{IncrementsRelative}}}

\item{increments}{see \code{\linkS4class{IncrementsRelative}}}
}
\value{
the \code{\linkS4class{IncrementsRelative}} object
}
\description{
Initialization function for "IncrementsRelative"
}
\keyword{methods}
