% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\docType{class}
\name{IncrementsRelativeSafety-class}
\alias{IncrementsRelativeSafety-class}
\alias{.IncrementsRelativeSafety}
\title{Increments control based on relative safety}
\description{
Increments control based on relative safety
}
\section{Slots}{

\describe{
\item{\code{intervals}}{a vector with the left bounds of the relevant intervals}

\item{\code{increments}}{a vector of the same length with the maximum allowable}

\item{\code{target}}{the target toxicity}

\item{\code{toxprob}}{the threshold probability that the dose is above the target}
}}

\keyword{classes}
