% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\name{IncrementsRelativeSafety}
\alias{IncrementsRelativeSafety}
\title{Initialization function for "IncrementsRelativeSafety"}
\usage{
IncrementsRelativeSafety(intervals, increments, target, toxprob)
}
\arguments{
\item{intervals}{see \code{\linkS4class{IncrementsRelativeSafety}}}

\item{increments}{see \code{\linkS4class{IncrementsRelativeSafety}}}

\item{target}{see \code{\linkS4class{IncrementsRelativeSafety}}}

\item{toxprob}{see \code{\linkS4class{IncrementsRelativeSafety}}}
}
\description{
Initialization function for "IncrementsRelativeSafety"
}
\keyword{methods}
