% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\docType{class}
\name{IncrementsRelativeSafetyFix-class}
\alias{IncrementsRelativeSafetyFix-class}
\alias{.IncrementsRelativeSafetyFix}
\title{Increments control based on relative safety}
\description{
Increments control based on relative safety
}
\section{Slots}{

\describe{
\item{\code{intervals}}{a vector with the left bounds of the relevant intervals}

\item{\code{increments}}{a vector of the same length with the maximum allowable}

\item{\code{toxrule}}{the toxrule for lower doses only}

\item{\code{toxrulesame}}{toxrule for lower and the same dose dose}
}}

\keyword{classes}
