% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\name{IncrementsRelativeSafetyFix}
\alias{IncrementsRelativeSafetyFix}
\title{Initialization function for "IncrementsRelativeSafetyFix"}
\usage{
IncrementsRelativeSafetyFix(intervals, increments, toxrule, toxrulesame)
}
\arguments{
\item{intervals}{see \code{\linkS4class{IncrementsRelativeSafetyFix}}}

\item{increments}{see \code{\linkS4class{IncrementsRelativeSafetyFix}}}

\item{toxrule}{see \code{\linkS4class{IncrementsRelativeSafetyFix}}}

\item{toxrulesame}{see \code{\linkS4class{IncrementsRelativeSafetyFix}}}
}
\description{
Initialization function for "IncrementsRelativeSafetyFix"
}
\keyword{methods}
