% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_IncrementsSafetyStop.R
\docType{class}
\name{IncrementsSafetyStop-class}
\alias{IncrementsSafetyStop-class}
\alias{.IncrementsSafetyStop}
\title{Increments control based on the number of observed DLTs
This is a safety rule that limits further escalation based on the
number of observed DLTs. The probability that a dose is toxic is calculated
using a Bin-Beta distribution with prior (a,b). If the probability exceeds
the threshold for a given dose, that dose and all doses above are excluded
from further escalation.}
\description{
Increments control based on the number of observed DLTs
This is a safety rule that limits further escalation based on the
number of observed DLTs. The probability that a dose is toxic is calculated
using a Bin-Beta distribution with prior (a,b). If the probability exceeds
the threshold for a given dose, that dose and all doses above are excluded
from further escalation.
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity}

\item{\code{prob}}{the threshold probability}

\item{\code{prior.a}}{the the first parameter of the prior Beta distribution}

\item{\code{prior.b}}{the the second parameter of the prior Beta distribution}
}}

\keyword{classes}
