% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_IncrementsSafetyStop.R
\name{IncrementsSafetyStop}
\alias{IncrementsSafetyStop}
\title{Initialization function for "IncrementsSafetyStop"}
\usage{
IncrementsSafetyStop(target, prob, prior.a = 1, prior.b = 1)
}
\arguments{
\item{target}{see \code{\linkS4class{IncrementsSafetyStop}}}

\item{prob}{see \code{\linkS4class{IncrementsSafetyStop}}}

\item{prior.a}{see \code{\linkS4class{IncrementsSafetyStop}}}

\item{prior.b}{see \code{\linkS4class{IncrementsSafetyStop}}}
}
\description{
Initialization function for "IncrementsSafetyStop"
}
\keyword{methods}
