% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_IncrementsSafetyStopFix.R
\docType{class}
\name{IncrementsSafetyStopFix-class}
\alias{IncrementsSafetyStopFix-class}
\alias{.IncrementsSafetyStopFix}
\title{Increments control based on the number of observed DLTs
This is a safety rule that uses fixed numbers for observed DLTs
at a dose level to limit future escalation. I.e. In case that the number of DLTs
exceeds a predefined number specified in the protocol only lower doses or the
same dose are allowed for the further dose escalation.}
\description{
Note: in case that samedose=TRUE, only the exact match of DLT is evaluated
i.e. the function returns ok, if the number of observed DLTs are above
the specified number. In this case the rule should only be used together
with a rule using samedose=FALSE. The rule must be specified as a
matrix with 2 rows.
}
\section{Slots}{

\describe{
\item{\code{toxrule}}{the toxrule for lower or the same dose. The rule must be given in a matrix form:
row1=Number of patients,
row2=number of DLTs}

\item{\code{samedose}}{If true the same dose is ok}
}}

\keyword{classes}
