% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_IncrementsSafetyStopFix.R
\name{IncrementsSafetyStopFix}
\alias{IncrementsSafetyStopFix}
\title{Initialization function for "IncrementsSafetyStopFix"}
\usage{
IncrementsSafetyStopFix(toxrule, samedose = FALSE)
}
\arguments{
\item{toxrule}{see \code{\linkS4class{IncrementsSafetyStopFix}}}

\item{samedose}{see \code{\linkS4class{IncrementsSafetyStopFix}}}
}
\description{
Initialization function for "IncrementsSafetyStopFix"
}
\keyword{methods}
