% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_LogisticKadane2_DK.R
\docType{class}
\name{LogisticKadane2-class}
\alias{LogisticKadane2-class}
\alias{.LogisticKadane2}
\title{Reparametrized logistic model (version 2)}
\description{
This is the logistic model in the parametrization (version 2) of
Kadane et al. (1980).
}
\details{
Let \eqn{\rho_{0} = p(x_{0})} be the probability of a DLT of the placebo
(no drug) dose \eqn{x_{0}}, and let \eqn{MTD} be the dose with target
toxicity probability \eqn{\theta}, i.e. \eqn{p(MTD) = \theta}. Then it can
easily be shown that the logistic regression model has
intercept
\deqn{\frac{MTD logit(\rho_{0})}{MTD}}
and slope
\deqn{\frac{logit(\theta) - logit(\rho_{0})}{MTD}}

The prior for \eqn{MTD} is a \eqn{Gamma(shape,rate)} distribution.
The prior for \eqn{\rho_{0}} is a \eqn{Beta(\alpha,\beta)} distribution.

The minimum \eqn{d_{min}} and maximum \eqn{d_{max}} planned dose, are used
to set the initial value of the \eqn{MTD} arbitrarily as the average of
those two.
The initial value of \eqn{\rho_{0}} is set arbitrarily as \deqn{\frac{\theta}{10}}.

The slots of this class, required for creating the model, are the target
toxicity, the Beta and Gamma distribution parameters, as well as the minimum
and maximum of the dose range. Note that these can be different from the
minimum and maximum of the dose grid in the data later on.
}
\section{Slots}{

\describe{
\item{\code{theta}}{the target toxicity probability \eqn{\theta}}

\item{\code{dmin}}{the minimum of the dose range \eqn{d_{min}}}

\item{\code{dmax}}{the maximum of the dose range \eqn{d_{max}}}

\item{\code{alpha}}{the \eqn{\alpha} shape parameter of the \eqn{Beta(\alpha,\beta)}
distribution}

\item{\code{beta}}{the \eqn{\beta} shape parameter of the \eqn{Beta(\alpha,\beta)}
distribution}

\item{\code{shape}}{the shape parameter of the \eqn{Gamma(shape,rate)} distribution}

\item{\code{rate}}{the rate parameter of the  \eqn{Gamma(shape,rate)} distribution.}
}}

\keyword{classes}
