% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_LogisticKadane2_DK.R
\name{LogisticKadane2}
\alias{LogisticKadane2}
\title{Initialization function for the "LogisticKadane2" class}
\usage{
LogisticKadane2(theta, dmin, dmax, alpha, beta, shape, rate)
}
\arguments{
\item{theta}{the target toxicity probability}

\item{dmin}{the minimum of the dose range}

\item{dmax}{the maximum of the dose range}

\item{alpha}{the \eqn{\alpha} shape parameter of the \eqn{Beta(\alpha,\beta)}
distribution}

\item{beta}{the \eqn{\beta} shape parameter of the \eqn{Beta(\alpha,\beta)}
distribution}

\item{shape}{the shape parameter of the \eqn{Gamma(shape,rate)} distribution}

\item{rate}{the rate parameter of the  \eqn{Gamma(shape,rate)} distribution.}
}
\value{
the \code{\linkS4class{LogisticKadane2}}
}
\description{
Initialization function for the "LogisticKadane2" class
}
\keyword{methods}
