% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_LogisticKadaneBayer.R
\docType{class}
\name{LogisticKadaneBayer-class}
\alias{LogisticKadaneBayer-class}
\alias{.LogisticKadaneBayer}
\title{Standard Kadane model}
\description{
Standard Kadane model
}
\section{Slots}{

\describe{
\item{\code{theta}}{theta}

\item{\code{dmin}}{dose min}

\item{\code{dmax}}{dose max}

\item{\code{alpha}}{intercept}

\item{\code{beta}}{slope}

\item{\code{shape}}{shape}

\item{\code{rate}}{rate}
}}

\keyword{classes}
