% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_LogisticKadaneBayer.R
\name{LogisticKadaneBayer}
\alias{LogisticKadaneBayer}
\title{Initialization function for the "LogisticKadaneBayer" class}
\usage{
LogisticKadaneBayer(theta, dmin, dmax, alpha, beta, shape, rate)
}
\arguments{
\item{theta}{theta}

\item{dmin}{dose min}

\item{dmax}{dose max}

\item{alpha}{intercept}

\item{beta}{slope}

\item{shape}{shape}

\item{rate}{rate}
}
\value{
the \code{\linkS4class{LogisticKadaneBayer}} object
}
\description{
Initialization function for the "LogisticKadaneBayer" class
}
\keyword{methods}
