% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\name{NextBestKadaneBay}
\alias{NextBestKadaneBay}
\title{Initialization function for class "NextBestKadaneBay"}
\usage{
NextBestKadaneBay(target)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestKadaneBay}}}
}
\description{
Initialization function for class "NextBestKadaneBay"
}
\keyword{methods}
