% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R, R/y_bay_rules_NextBestMinDist.R
\docType{class}
\name{NextBestMinDist-class}
\alias{NextBestMinDist-class}
\alias{.NextBestMinDist}
\alias{NextBestMinDist}
\title{Class for Next Best Dose based on Minimum Distance to Target Probability}
\usage{
NextBestMinDist(target)
}
\arguments{
\item{target}{target DLT probability.}
}
\description{
This is typically used for CRM designs where a single target DLT probability
is targeted.
}
\section{Functions}{
\itemize{
\item \code{NextBestMinDist}: Initialization function for \code{NextBestMinDist}.
}}

\section{Slots}{

\describe{
\item{\code{target}}{target DLT probability.}

\item{\code{target}}{the target toxicity probability}
}}

\keyword{classes}
