% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_NextBestMinDist.R
\name{NextBestMinDist}
\alias{NextBestMinDist}
\title{Initialization function for class "NextBestMinDist"}
\usage{
NextBestMinDist(target)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestMTD}}}
}
\description{
Initialization function for class "NextBestMinDist"
}
\keyword{methods}
