% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_NextBestMTDCRM_DK.R
\docType{class}
\name{NextBestMTDCRM-class}
\alias{NextBestMTDCRM-class}
\alias{.NextBestMTDCRM}
\title{The class with the input for finding the next best dose based on CRM MTD
allocation probability.}
\description{
The class with the input for finding the next best dose based on CRM MTD
allocation probability.
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity probability}

\item{\code{pbomethod}}{the method used to handle the allocation probability for the
zero dose. Applicable to cases where data@plabebo = "FALSE"
If \code{none}, then the number of simulation instances allocated to zero
dose, are exuded from the derivation of allocation probabilities.
If \code{min}, then the number of simulation instances allocated to zero
dose, are combined with the instances allocated to the minimum planned dose.
If \code{max}, then the number of simulation instances allocated to zero
dose, are combined with the instances allocated to the maximum planned dose.}
}}

\keyword{classes}
