% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_NextBestMTDCRM_DK.R
\name{NextBestMTDCRM}
\alias{NextBestMTDCRM}
\title{Initialization function for class "NextBestMTDCRM"}
\usage{
NextBestMTDCRM(target, pbomethod = c("none", "min", "max"))
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestMTDCRM}}}

\item{pbomethod}{see \code{\linkS4class{NextBestMTDCRM}}}
}
\value{
the \code{\linkS4class{NextBestMTDCRM}} object
}
\description{
Initialization function for class "NextBestMTDCRM"
}
\keyword{methods}
