% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_NextBestMTDprob.R
\docType{class}
\name{NextBestMTDprob-class}
\alias{NextBestMTDprob-class}
\alias{.NextBestMTDprob}
\title{The class with the input for finding the next best MTD estimate with highest
probability to be the MTD.
The dose which has the highest probability to be the MTD among the specified
doses is used as next best dose. The posterior is calculated for each
iteration and the maximum dose that is not toxic (below target) is determined.
The allocation criteria is calculated as the frequency of a dose being
selected as MTD. The dose with the highest allocation value is selected as
next best dose.}
\description{
The class with the input for finding the next best MTD estimate with highest
probability to be the MTD.
The dose which has the highest probability to be the MTD among the specified
doses is used as next best dose. The posterior is calculated for each
iteration and the maximum dose that is not toxic (below target) is determined.
The allocation criteria is calculated as the frequency of a dose being
selected as MTD. The dose with the highest allocation value is selected as
next best dose.
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity probability}

\item{\code{method}}{the method used for calculating the allocation criteria
1=aggregate no dose safe with first dose
2=aggregate no dose safe with last dose
3=do not aggregate, but report the first dose as best next dose, in case 'no
dose is safe' has the highest allocation value. This follows the idea used for
the smallest distance to the target dose approach, where always a dose is
recommended, even though the lowest dose estimated is above target}
}}

\keyword{classes}
