% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_NextBestMTDprob.R
\name{NextBestMTDprob}
\alias{NextBestMTDprob}
\title{Initialization function for class "NextBestMTDprob"}
\usage{
NextBestMTDprob(target, method = 1)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestMTDprob}}}

\item{method}{see \code{\linkS4class{NextBestMTDprob}}}
}
\description{
Initialization function for class "NextBestMTDprob"
}
\keyword{methods}
