% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\docType{class}
\name{NextBestTwoParBay-class}
\alias{NextBestTwoParBay-class}
\alias{.NextBestTwoParBay}
\title{The class with the input for finding the next best MTD estimate with highest
probability to be the MTD}
\description{
The class with the input for finding the next best MTD estimate with highest
probability to be the MTD
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity probability}
}}

\keyword{classes}
