% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\name{NextBestTwoParBay}
\alias{NextBestTwoParBay}
\title{Initialization function for class "NextBestTwoParBay"}
\usage{
NextBestTwoParBay(target)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestTwoParBay}}}
}
\description{
Initialization function for class "NextBestTwoParBay"
}
\keyword{methods}
