% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_OneParExp.R
\docType{class}
\name{OneParExp-class}
\alias{OneParExp-class}
\alias{.OneParExp}
\title{Standard one parameter model}
\description{
Standard one parameter model
}
\section{Slots}{

\describe{
\item{\code{skeletonFun}}{the skeleton}

\item{\code{skeletonProbs}}{the skeleton probs}

\item{\code{lambda}}{the parameter lambda}
}}

\keyword{classes}
