% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_OneParExp.R
\name{OneParExp}
\alias{OneParExp}
\title{Initialization function for the "OneParExp" class}
\usage{
OneParExp(skeletonProbs, doseGrid, lambda)
}
\arguments{
\item{skeletonProbs}{the skeleton probs}

\item{doseGrid}{the dose grid}

\item{lambda}{the lambda}
}
\value{
the \code{\linkS4class{OneParExp}} object
}
\description{
Initialization function for the "OneParExp" class
}
\keyword{methods}
