% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_design_2GeneralSimulationsBay_show.R
\name{show,GeneralSimulationsSummaryBay-method}
\alias{show,GeneralSimulationsSummaryBay-method}
\title{Show the summary of the simulations}
\usage{
\S4method{show}{GeneralSimulationsSummaryBay}(object)
}
\arguments{
\item{object}{the \code{\linkS4class{GeneralSimulationsSummaryBay}} object we want
to print}
}
\value{
invisibly returns a data frame of the results with one row and
appropriate column names
}
\description{
Show the summary of the simulations
}
\keyword{methods}
