% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_design_3SimulationsBay_summary.R
\name{SimulationsBay}
\alias{SimulationsBay}
\title{Initialization function for the "Simulations" class}
\usage{
SimulationsBay(fit, estimates, stopReasons, ...)
}
\arguments{
\item{fit}{see \code{\linkS4class{Simulations}}}

\item{estimates}{see \code{\linkS4class{Simulations}}}

\item{stopReasons}{see \code{\linkS4class{Simulations}}}

\item{\dots}{additional parameters from \code{\link{GeneralSimulations}}}
}
\value{
the \code{\linkS4class{Simulations}} object
}
\description{
Initialization function for the "Simulations" class
}
\keyword{methods}
