% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingHardFirst.R
\docType{class}
\name{StoppingHardFirst-class}
\alias{StoppingHardFirst-class}
\alias{.StoppingHardFirst}
\title{Hardstopping rule based on the Bin-Beta distribution and not on the
results from the model.The rule is triggered when
the first dose is considered to be toxic (above threshold prob) based on the
observed data and a Beta prior distribution. For calculation the posterior
bin-beta distribution, usually an uninformative Beta(1,1) prior is used.}
\description{
Hardstopping rule based on the Bin-Beta distribution and not on the
results from the model.The rule is triggered when
the first dose is considered to be toxic (above threshold prob) based on the
observed data and a Beta prior distribution. For calculation the posterior
bin-beta distribution, usually an uninformative Beta(1,1) prior is used.
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity}

\item{\code{prob}}{the threshold probability}

\item{\code{prior.a}}{the first parameter of the prior Beta distribution}

\item{\code{prior.b}}{the second parameter of the prior Beta distribution}

\item{\code{label}}{label of stopping rule}
}}

\keyword{classes}
