% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingHardFirst.R
\name{StoppingHardFirst}
\alias{StoppingHardFirst}
\title{Initialization function for "StoppingHardFirst"}
\usage{
StoppingHardFirst(target, prob, prior.a = 1, prior.b = 1, label)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingHardFirst}}}

\item{prob}{see \code{\linkS4class{StoppingHardFirst}}}

\item{prior.a}{see \code{\linkS4class{StoppingHardFirst}}}

\item{prior.b}{see \code{\linkS4class{StoppingHardFirst}}}

\item{label}{see \code{\linkS4class{StoppingHardFirst}}}
}
\description{
Initialization function for "StoppingHardFirst"
}
\keyword{methods}
