% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingHardFirstFix.R
\docType{class}
\name{StoppingHardFirstFix-class}
\alias{StoppingHardFirstFix-class}
\alias{.StoppingHardFirstFix}
\title{Stopping rule based on the number of observed DLTs at a dose level.
This is a hards stopping rule. The unacceptable numer of DLTs are
specified in the protocol. I.e. in the protocol a certain numbers for
stopping the study is specified
(often not based on a posterior distribution)}
\description{
Stopping rule based on the number of observed DLTs at a dose level.
This is a hards stopping rule. The unacceptable numer of DLTs are
specified in the protocol. I.e. in the protocol a certain numbers for
stopping the study is specified
(often not based on a posterior distribution)
}
\section{Slots}{

\describe{
\item{\code{toxrule}}{matrix with toxrule}

\item{\code{label}}{label of stopping rule}
}}

\keyword{classes}
