% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingHardFirstFix.R
\name{StoppingHardFirstFix}
\alias{StoppingHardFirstFix}
\title{Initialization function for "StoppingHardFirstFix"}
\usage{
StoppingHardFirstFix(toxrule, label)
}
\arguments{
\item{toxrule}{see \code{\linkS4class{StoppingHardFirstFix}}}

\item{label}{see \code{\linkS4class{StoppingHardFirstFix}}}
}
\description{
Initialization function for "StoppingHardFirstFix"
}
\keyword{methods}
