% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_StoppingLowestDoseHSRBeta_DK.R
\docType{class}
\name{StoppingLowestDoseHSRBeta_DK-class}
\alias{StoppingLowestDoseHSRBeta_DK-class}
\alias{.StoppingLowestDoseHSRBeta_DK}
\title{Stop based on the lowest dose meeting the hard safety criteria using
Beta based DLT probability, i.e. 1- pbeta(target, x+a, n-x+b) < prob}
\description{
Stop based on the lowest dose meeting the hard safety criteria using
Beta based DLT probability, i.e. 1- pbeta(target, x+a, n-x+b) < prob
}
\section{Slots}{

\describe{
\item{\code{target}}{toxicity target for a dose}

\item{\code{prob}}{probability of dose being toxic}

\item{\code{a}}{shape parameter a>0 of probability distribution Beta (a,b)}

\item{\code{b}}{shape parameter b>0 of probability distribution Beta (a,b)}
}}

\keyword{classes}
