% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_StoppingLowestDoseHSRBeta_DK.R
\name{StoppingLowestDoseHSRBeta_DK}
\alias{StoppingLowestDoseHSRBeta_DK}
\title{Initialization function for "StoppingLowestDoseHSRBeta_DK"}
\usage{
StoppingLowestDoseHSRBeta_DK(target, prob, a, b)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingLowestDoseHSRBeta_DK}}}

\item{prob}{see \code{\linkS4class{StoppingLowestDoseHSRBeta_DK}}}

\item{a}{see \code{\linkS4class{StoppingLowestDoseHSRBeta_DK}}}

\item{b}{see \code{\linkS4class{StoppingLowestDoseHSRBeta_DK}}}
}
\value{
the \code{\linkS4class{StoppingLowestDoseHSRBeta_DK}} object
}
\description{
Initialization function for "StoppingLowestDoseHSRBeta_DK"
}
\keyword{methods}
