% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingMTDdistributionBay.R
\docType{class}
\name{StoppingMTDdistributionBay-class}
\alias{StoppingMTDdistributionBay-class}
\alias{.StoppingMTDdistributionBay}
\title{stopping rule based on the posterior distribution}
\description{
stopping rule based on the posterior distribution
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity}

\item{\code{prob}}{the threshold probability}

\item{\code{doseeval}}{the dose were the rule is evaluated}

\item{\code{direction}}{the direction, is the dose toxic or safe}

\item{\code{dosetested}}{must the dose be already tested}

\item{\code{label}}{label of stopping rule}
}}

\keyword{classes}
