% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingMTDdistributionBay.R
\name{StoppingMTDdistributionBay}
\alias{StoppingMTDdistributionBay}
\title{Initialization function for "StoppingMTDdistributionBay"}
\usage{
StoppingMTDdistributionBay(
  target,
  prob,
  doseeval,
  direction,
  dosetested = TRUE,
  label
)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingMTDdistributionBay}}}

\item{prob}{see \code{\linkS4class{StoppingMTDdistributionBay}}}

\item{doseeval}{see \code{\linkS4class{StoppingMTDdistributionBay}}}

\item{direction}{see \code{\linkS4class{StoppingMTDdistributionBay}}}

\item{dosetested}{see \code{\linkS4class{StoppingMTDdistributionBay}}}

\item{label}{see \code{\linkS4class{StoppingMTDdistributionBay}}}
}
\description{
Initialization function for "StoppingMTDdistributionBay"
}
\keyword{methods}
