% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingPatientsNearDose2.R
\docType{class}
\name{StoppingPatientsNearDose2-class}
\alias{StoppingPatientsNearDose2-class}
\alias{.StoppingPatientsNearDose2}
\title{Stop based on number of patients near to next best dose
!! No rule change to standard StoppingPatientsNearDose
Only change is returned text, when h stopping rule is triggered}
\description{
Stop based on number of patients near to next best dose
!! No rule change to standard StoppingPatientsNearDose
Only change is returned text, when h stopping rule is triggered
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{number of required patients}

\item{\code{percentage}}{percentage (between 0 and 100) within the next best dose
the patients must lie}

\item{\code{label}}{label of stopping rule}
}}

\examples{

# As example, here is the rule for: 
#   stopping the study if at least 9 patients were dosed at a dose within (1 +/- 0.2) 
#   of the next best dose

myStopping <- StoppingPatientsNearDose(nPatients = 9,
                                       percentage = 0.2)

}
\keyword{classes}
