% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingPatientsNearDose2.R
\name{StoppingPatientsNearDose2}
\alias{StoppingPatientsNearDose2}
\title{Initialization function for "StoppingPatientsNearDose2"}
\usage{
StoppingPatientsNearDose2(nPatients, percentage, label)
}
\arguments{
\item{nPatients}{see \code{\linkS4class{StoppingPatientsNearDose}}}

\item{percentage}{see \code{\linkS4class{StoppingPatientsNearDose}}}

\item{label}{see \code{\linkS4class{StoppingPatientsNearDose}}}
}
\value{
the \code{\linkS4class{StoppingPatientsNearDose}} object
}
\description{
Initialization function for "StoppingPatientsNearDose2"
}
\keyword{methods}
