% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_myfunction-outdated.R
\name{StoppingPatientsNearDoseBay}
\alias{StoppingPatientsNearDoseBay}
\title{Initialization function for "StoppingPatientsNearDose"}
\usage{
StoppingPatientsNearDoseBay(nPatients, percentage)
}
\arguments{
\item{nPatients}{see \code{\linkS4class{StoppingPatientsNearDose}}}

\item{percentage}{see \code{\linkS4class{StoppingPatientsNearDose}}}
}
\value{
the \code{\linkS4class{StoppingPatientsNearDose}} object
}
\description{
Initialization function for "StoppingPatientsNearDose"
}
\keyword{methods}
