% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingProbBay.R
\docType{class}
\name{StoppingProbBay-class}
\alias{StoppingProbBay-class}
\alias{.StoppingProbBay}
\title{Stopping rule based on the posterior probability of the full model.
The probability that a certain dose is toxic or safe will be calculated
and compared to a threshold. The study is stopped in case that the
probability is below or above the threshold (toxic or safe).}
\description{
Stopping rule based on the posterior probability of the full model.
The probability that a certain dose is toxic or safe will be calculated
and compared to a threshold. The study is stopped in case that the
probability is below or above the threshold (toxic or safe).
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity}

\item{\code{prob}}{the threshold probability}

\item{\code{doseeval}}{the dose were the rule is evaluated}

\item{\code{direction}}{the direction, is the dose toxic or safe}

\item{\code{dosetested}}{must the dose be already tested}

\item{\code{label}}{label of stopping rule}
}}

\keyword{classes}
