% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_rules_StoppingProbBay.R
\name{StoppingProbBay}
\alias{StoppingProbBay}
\title{Initialization function for "StoppingProbBay"}
\usage{
StoppingProbBay(target, prob, doseeval, direction, dosetested = TRUE, label)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingProbBay}}}

\item{prob}{see \code{\linkS4class{StoppingProbBay}}}

\item{doseeval}{see \code{\linkS4class{StoppingProbBay}}}

\item{direction}{see \code{\linkS4class{StoppingProbBay}}}

\item{dosetested}{see \code{\linkS4class{StoppingProbBay}}}

\item{label}{see \code{\linkS4class{StoppingProbBay}}}
}
\description{
Initialization function for "StoppingProbBay"
}
\keyword{methods}
