% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ZZ_StoppingTargetProbPatientsNearLowestDose_DK.R
\name{StoppingTargetProbPatientsNearLowestDose}
\alias{StoppingTargetProbPatientsNearLowestDose}
\title{Initialization function for "StoppingTargetProbPatientsNearLowestDose"}
\usage{
StoppingTargetProbPatientsNearLowestDose(target, prob, nPatients, percentage)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingTargetProb}}}

\item{prob}{see \code{\linkS4class{StoppingTargetProb}}}

\item{nPatients}{see \code{\linkS4class{StoppingPatientsNearDose}}}

\item{percentage}{see \code{\linkS4class{StoppingPatientsNearDose}}}
}
\value{
the \code{\linkS4class{StoppingTargetProb}} object

the \code{\linkS4class{StoppingPatientsNearDose}} object
}
\description{
Initialization function for "StoppingTargetProbPatientsNearLowestDose"
}
\keyword{methods}
