% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_design_1GeneralSimulationsBay_summary.R
\name{summary,GeneralSimulationsBay-method}
\alias{summary,GeneralSimulationsBay-method}
\title{Summarize the simulations, relative to a given truth}
\usage{
\S4method{summary}{GeneralSimulationsBay}(object, truth, target = 0.3, ...)
}
\arguments{
\item{object}{the \code{\linkS4class{GeneralSimulations}} object we want to
summarize}

\item{truth}{a function which takes as input a dose (vector) and returns the
true probability (vector) for toxicity}

\item{target}{the target toxicity interval (default: 20-35\\%) used for the
computations}

\item{\dots}{Additional arguments can be supplied here for \code{truth}}
}
\value{
an object of class \code{\linkS4class{GeneralSimulationsSummary}}
}
\description{
Summarize the simulations, relative to a given truth
}
\keyword{methods}
