% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_TwoPar.R
\docType{class}
\name{TwoParBay-class}
\alias{TwoParBay-class}
\alias{.TwoParBay}
\title{Standard two parameter model}
\description{
Standard two parameter model
}
\section{Slots}{

\describe{
\item{\code{mean1}}{the mean of the intercept}

\item{\code{mean2}}{the mean of the slope}

\item{\code{var1}}{the variance of the intercept}

\item{\code{var2}}{the variance of the slope}
}}

\keyword{classes}
