% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_TwoPar.R
\name{TwoParBay}
\alias{TwoParBay}
\title{Initialization function for the "TwoParBay" class}
\usage{
TwoParBay(mean1, mean2, var1, var2)
}
\arguments{
\item{mean1}{the mean of the intercept}

\item{mean2}{the mean of the slope}

\item{var1}{the variance of the intercept}

\item{var2}{the variance of the slope}
}
\value{
the \code{\linkS4class{TwoParBay}} object
}
\description{
Initialization function for the "TwoParBay" class
}
\keyword{methods}
