% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_TwoParBayS.R
\docType{class}
\name{TwoParBayS-class}
\alias{TwoParBayS-class}
\alias{.TwoParBayS}
\title{Implementation of the two parameter logistic model without scaling
factor 100}
\description{
Implementation of the two parameter logistic model without scaling
factor 100
}
\section{Slots}{

\describe{
\item{\code{mean1}}{the mean of the intercept}

\item{\code{mean2}}{the mean of the slope}

\item{\code{var1}}{the variance of the intercept}

\item{\code{var2}}{the variance of the slope}
}}

\keyword{classes}
