% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_bay_model_TwoParBayS.R
\name{TwoParBayS}
\alias{TwoParBayS}
\title{Initialization function for the "TwoParBayS" class}
\usage{
TwoParBayS(mean1, mean2, var1, var2)
}
\arguments{
\item{mean1}{the mean of the intercept}

\item{mean2}{the mean of the slope}

\item{var1}{the variance of the intercept}

\item{var2}{the variance of the slope}
}
\value{
the \code{\linkS4class{TwoParBayS}} object
}
\description{
Initialization function for the "TwoParBayS" class
}
\keyword{methods}
